+++ srcs/ms_parse.c
@@ -3,14 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ms_parse.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sthitiku <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: nnakarac <nnakarac@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/20 01:04:35 by sthitiku          #+#    #+#             */
-/*   Updated: 2023/01/20 00:28:47 by sthitiku         ###   ########.fr       */
+/*   Updated: 2023/01/23 09:49:41 by nnakarac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ms_input.h"
+
+extern int	g_status;
 
 static char	*cut_quote(char *str, char quote)
 {
@@ -61,7 +63,11 @@ static char	*parse_env_get_env(char *str, int str_len, char first, char last)
 	else
 	{
 		p.sub = ft_substr(str, p.start + 1, str_len - p.start - 1);
-		p.env = getenv(p.sub);
+		// get previous exit status "$?"
+		if (str_len == 2 && !ft_strncmp(p.sub, "?", 1))
+			p.env = ft_itoa(g_status);
+		else
+			p.env = getenv(p.sub);
 		new = malloc(sizeof(char) * ft_strlen(p.env) + 1);
 		ft_strlcpy(new, p.env, ft_strlen(p.env) + 1);
 	}
